requirements.txt
# Utilities
numpy
pandas

# FACE RECOGNITION
scikit-learn

dlib-bin
git+https://github.com/ageitgey/face_recognition_models
setuptools<70.0.0

supabase
bcrypt

segno
pillow

streamlit

librosa
resemblyzer